AI_Detection_Engine/requirements.txt
opencv-python
torch
transformers
pillow
numpy
scikit-learn
streamlit
